requirements.txt
streamlit>=1.31.0
python-dotenv>=1.0.0
pypdf>=4.0.1
python-docx>=1.1.0
faiss-cpu>=1.7.4
numpy>=1.26.0
ollama>=0.1.9
google-generativeai>=0.3.0
requests>=2.31.0
